# Improve decoding performance by memoizing higher bits