--- a/decode_generic.h
+++ b/decode_generic.h
@@ -131,7 +131,7 @@ static inline void set_ones_at_bit_pos(WIDTH)(TYPE *values,
 }
 
 static inline void bcltree_add(WIDTH)(struct DecodeCtx(WIDTH) *ctx,
-  const struct BitCluster *cluster)
+  const struct decode_bit_cluster *cluster)
 {
   if (cluster->bit_pos == 0)
     return;
@@ -147,17 +147,17 @@ static inline int bcltree_has_more(WIDTH)(struct DecodeCtx(WIDTH) *ctx)
   return !decode_stack_empty(&ctx->stack);
 }
 
-static inline struct BitCluster *bcltree_next(WIDTH)(struct DecodeCtx(WIDTH) *ctx)
+static inline struct decode_bit_cluster *bcltree_next(WIDTH)(struct DecodeCtx(WIDTH) *ctx)
 {
   return decode_stack_pop(&ctx->stack);
 }
 
 static VtencErrorCode decode_bit_cluster_tree(WIDTH)(struct DecodeCtx(WIDTH) *ctx)
 {
-  bcltree_add(WIDTH)(ctx, &(struct BitCluster){0, ctx->values_len, WIDTH});
+  bcltree_add(WIDTH)(ctx, &(struct decode_bit_cluster){0, ctx->values_len, WIDTH, 0});
 
   while (bcltree_has_more(WIDTH)(ctx)) {
-    struct BitCluster *cluster = bcltree_next(WIDTH)(ctx);
+    struct decode_bit_cluster *cluster = bcltree_next(WIDTH)(ctx);
     size_t cl_from = cluster->from;
     size_t cl_len = cluster->length;
     unsigned int cl_bit_pos = cluster->bit_pos;
@@ -183,8 +183,8 @@ static VtencErrorCode decode_bit_cluster_tree(WIDTH)(struct DecodeCtx(WIDTH) *ct
     set_ones_at_bit_pos(WIDTH)(ctx->values + cl_from + n_zeros, cl_len - n_zeros, cur_bit_pos);
 
     {
-      struct BitCluster zeros_cluster = {cl_from, n_zeros, cur_bit_pos};
-      struct BitCluster ones_cluster = {cl_from + n_zeros, cl_len - n_zeros, cur_bit_pos};
+      struct decode_bit_cluster zeros_cluster = {cl_from, n_zeros, cur_bit_pos};
+      struct decode_bit_cluster ones_cluster = {cl_from + n_zeros, cl_len - n_zeros, cur_bit_pos};
 
       bcltree_add(WIDTH)(ctx, &ones_cluster);
       bcltree_add(WIDTH)(ctx, &zeros_cluster);