--- a/encode_generic.h
+++ b/encode_generic.h
@@ -108,7 +108,7 @@ static inline VtencErrorCode encode_lower_bits(WIDTH)(struct EncodeCtx(WIDTH) *c
 }
 
 static inline void bcltree_add(WIDTH)(struct EncodeCtx(WIDTH) *ctx,
-  const struct encode_bit_cluster *cluster)
+  const struct enc_bit_cluster *cluster)
 {
   if (cluster->bit_pos == 0)
     return;
@@ -124,17 +124,17 @@ static inline int bcltree_has_more(WIDTH)(struct EncodeCtx(WIDTH) *ctx)
   return !enc_stack_empty(&ctx->stack);
 }
 
-static inline struct encode_bit_cluster *bcltree_next(WIDTH)(struct EncodeCtx(WIDTH) *ctx)
+static inline struct enc_bit_cluster *bcltree_next(WIDTH)(struct EncodeCtx(WIDTH) *ctx)
 {
   return enc_stack_pop(&ctx->stack);
 }
 
 static VtencErrorCode encode_bit_cluster_tree(WIDTH)(struct EncodeCtx(WIDTH) *ctx)
 {
-  bcltree_add(WIDTH)(ctx, &(struct encode_bit_cluster){0, ctx->values_len, WIDTH});
+  bcltree_add(WIDTH)(ctx, &(struct enc_bit_cluster){0, ctx->values_len, WIDTH});
 
   while (bcltree_has_more(WIDTH)(ctx)) {
-    struct encode_bit_cluster *cluster = bcltree_next(WIDTH)(ctx);
+    struct enc_bit_cluster *cluster = bcltree_next(WIDTH)(ctx);
     size_t cl_from = cluster->from;
     size_t cl_len = cluster->length;
     unsigned int cl_bit_pos = cluster->bit_pos;
@@ -153,8 +153,8 @@ static VtencErrorCode encode_bit_cluster_tree(WIDTH)(struct EncodeCtx(WIDTH) *ct
     RETURN_IF_ERROR(bswriter_write(&(ctx->bits_writer), n_zeros, enc_len));
 
     {
-      struct encode_bit_cluster zeros_cluster = {cl_from, n_zeros, cur_bit_pos};
-      struct encode_bit_cluster ones_cluster = {cl_from + n_zeros, cl_len - n_zeros, cur_bit_pos};
+      struct enc_bit_cluster zeros_cluster = {cl_from, n_zeros, cur_bit_pos};
+      struct enc_bit_cluster ones_cluster = {cl_from + n_zeros, cl_len - n_zeros, cur_bit_pos};
 
       bcltree_add(WIDTH)(ctx, &ones_cluster);
       bcltree_add(WIDTH)(ctx, &zeros_cluster);