When decoding, we reconstruct the list of values by calling the `set_ones_at_bit_pos` function on each node visited when traversing the Bit Cluster Tree. The function sets bits vertically, meaning that it has to move back and forth through the output list. This is inefficient as it causes unnecessary context switches.

This PR changes the way values are reconstructed. For each node seen on the Bit Cluster Tree, a value for the "higher bits" up to that node (or bit cluster) is calculated and stored. Children nodes use their parent's "higher bits" plus a `0` or `1` to calculate their own value. When a leaf node (`bit_pos` = 0) is hit, the decoding function simply uses the `higher_bits` value to append it to the output list. Adding values to the output is carried out in a linear way, from 0 through `out_len`.

Here's some stats on context switching before and after this change:

**BEFORE:**

```sh
$ sudo perf stat ./tests/gov2 ../data/integer/gov2.sorted 
number of sets processed: 1107205
input size: 23918861764 bytes
encoded size: 2885170530 bytes
compression ratio: 0.120623 (12.0623%)

 Performance counter stats for './tests/gov2 ../data/integer/gov2.sorted':

        398,704.44 msec task-clock                #    0.999 CPUs utilized          
             2,800      context-switches          #    0.007 K/sec                  
                41      cpu-migrations            #    0.000 K/sec                  
           918,794      page-faults               #    0.002 M/sec                  
 1,034,034,095,159      cycles                    #    2.593 GHz                    
 2,161,909,819,308      instructions              #    2.09  insn per cycle         
   412,957,054,242      branches                  # 1035.747 M/sec                  
    13,884,283,279      branch-misses             #    3.36% of all branches        

     398.928553975 seconds time elapsed

     388.550252000 seconds user
      10.156085000 seconds sys
```

**AFTER:**

```sh
$ sudo perf stat ./tests/gov2 ../data/integer/gov2.sorted 
number of sets processed: 1107205
input size: 23918861764 bytes
encoded size: 2885170530 bytes
compression ratio: 0.120623 (12.0623%)

 Performance counter stats for './tests/gov2 ../data/integer/gov2.sorted':

        357,963.84 msec task-clock                #    1.000 CPUs utilized          
             2,090      context-switches          #    0.006 K/sec                  
                29      cpu-migrations            #    0.000 K/sec                  
           918,796      page-faults               #    0.003 M/sec                  
   928,355,843,127      cycles                    #    2.593 GHz                    
 2,059,150,585,596      instructions              #    2.22  insn per cycle         
   382,110,669,249      branches                  # 1067.456 M/sec                  
    10,849,470,045      branch-misses             #    2.84% of all branches        

     357.957015376 seconds time elapsed

     347.216292000 seconds user
      10.748752000 seconds sys
```

It goes from 2,800 context-switches down to 2,090, a 25.36% less.